public/Cube/Scripts/scripts/magic-custom-cube-scripts/requirements.txt
Pillow>=9.0.0  # PIL (Python Imaging Library) for image processing
fpdf>=2.5.0    # For PDF creation
tqdm>=4.65.0   # For progress bars
